finished refactoring all four call methods to extend call types
