+++ lib/compose.js
@@ -1,6 +1,6 @@
 const generateUnaryCall = require("./calls/generateUnaryCall");
 const generateClientStreamCall = require("./calls/generateClientStreamCall");
-const ServerStreamCall = require("./calls/serverStreamCall");
+const generateServerStreamCall = require("./calls/generateServerStreamCall");
 const generateDuplex = require("./calls/generateDuplex");
 const getMethodType = require("./utils/getMethodType");
 
@@ -45,7 +45,7 @@ module.exports = function compose(
     case "ServerStream":
       // class Context {}
       return async function(call) {
-        let serverStreamCall = new ServerStreamCall(call);
+        let serverStreamCall = generateServerStreamCall(call);
         for (let i = 0; i < middlewareStack.length; i++) {
           await middlewareStack[i](serverStreamCall);
         }
